+++ src/main/java/org/opensearch/security/privileges/ResourceAccessEvaluator.java
@@ -11,7 +11,6 @@
 package org.opensearch.security.privileges;
 
 import java.util.List;
-import java.util.Set;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -20,11 +19,11 @@ import org.opensearch.action.ActionRequest;
 import org.opensearch.action.DocRequest;
 import org.opensearch.action.DocWriteRequest;
 import org.opensearch.action.get.GetRequest;
-import org.opensearch.common.settings.Settings;
 import org.opensearch.core.action.ActionListener;
 import org.opensearch.core.common.Strings;
 import org.opensearch.security.resources.ResourceAccessHandler;
-import org.opensearch.security.support.ConfigConstants;
+import org.opensearch.security.resources.ResourcePluginInfo;
+import org.opensearch.security.setting.OpensearchDynamicSetting;
 
 /**
  * Evaluates access to resources. The resource plugins must register the indices which hold resource information.
@@ -42,14 +41,22 @@ import org.opensearch.security.support.ConfigConstants;
 public class ResourceAccessEvaluator {
     private static final Logger log = LogManager.getLogger(ResourceAccessEvaluator.class);
 
-    private final Set<String> resourceIndices;
-    private final Settings settings;
+    private final ResourcePluginInfo resourcePluginInfo;
     private final ResourceAccessHandler resourceAccessHandler;
 
-    public ResourceAccessEvaluator(Set<String> resourceIndices, Settings settings, ResourceAccessHandler resourceAccessHandler) {
-        this.resourceIndices = resourceIndices;
-        this.settings = settings;
+    private final OpensearchDynamicSetting<Boolean> resourceSharingEnabledSetting;
+    private final OpensearchDynamicSetting<List<String>> protectedResourceTypesSetting;
+
+    public ResourceAccessEvaluator(
+        ResourcePluginInfo resourcePluginInfo,
+        ResourceAccessHandler resourceAccessHandler,
+        final OpensearchDynamicSetting<Boolean> resourceSharingEnabledSetting,
+        final OpensearchDynamicSetting<List<String>> protectedResourceTypesSetting
+    ) {
+        this.resourcePluginInfo = resourcePluginInfo;
         this.resourceAccessHandler = resourceAccessHandler;
+        this.resourceSharingEnabledSetting = resourceSharingEnabledSetting;
+        this.protectedResourceTypesSetting = protectedResourceTypesSetting;
     }
 
     /**
@@ -96,14 +103,9 @@ public class ResourceAccessEvaluator {
      * @return true if request should be evaluated, false otherwise
      */
     public boolean shouldEvaluate(ActionRequest request) {
-        boolean isResourceSharingFeatureEnabled = settings.getAsBoolean(
-            ConfigConstants.OPENSEARCH_RESOURCE_SHARING_ENABLED,
-            ConfigConstants.OPENSEARCH_RESOURCE_SHARING_ENABLED_DEFAULT
-        );
-        List<String> protectedTypes = settings.getAsList(
-            ConfigConstants.OPENSEARCH_RESOURCE_SHARING_PROTECTED_TYPES,
-            ConfigConstants.OPENSEARCH_RESOURCE_SHARING_PROTECTED_TYPES_DEFAULT
-        );
+        boolean isResourceSharingFeatureEnabled = resourceSharingEnabledSetting.getDynamicSettingValue();
+        List<String> protectedTypes = protectedResourceTypesSetting.getDynamicSettingValue();
+
         if (!isResourceSharingFeatureEnabled) return false;
         if (!(request instanceof DocRequest docRequest)) return false;
         /**
@@ -128,7 +130,7 @@ public class ResourceAccessEvaluator {
             return false;
         }
         // if requested index is not a resource sharing index, move on to the regular evaluator
-        if (!resourceIndices.contains(docRequest.index())) {
+        if (!resourcePluginInfo.getResourceIndicesForProtectedTypes().contains(docRequest.index())) {
             log.debug("Request index {} is not a protected resource index", docRequest.index());
             return false;
         }