+++ src/main/java/org/opensearch/security/resources/ResourceAccessControlClient.java
@@ -8,15 +8,15 @@
 
 package org.opensearch.security.resources;
 
+import java.util.List;
 import java.util.Set;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import org.opensearch.core.action.ActionListener;
+import org.opensearch.security.setting.OpensearchDynamicSetting;
 import org.opensearch.security.spi.resources.client.ResourceSharingClient;
-import org.opensearch.security.spi.resources.sharing.ResourceSharing;
-import org.opensearch.security.spi.resources.sharing.ShareWith;
 
 /**
  * Access control client responsible for handling resource sharing operations such as verifying,
@@ -28,15 +28,21 @@ public final class ResourceAccessControlClient implements ResourceSharingClient 
     private static final Logger LOGGER = LogManager.getLogger(ResourceAccessControlClient.class);
 
     private final ResourceAccessHandler resourceAccessHandler;
-    private final Set<String> resourceIndices;
+    private final ResourcePluginInfo resourcePluginInfo;
+    private final OpensearchDynamicSetting<List<String>> resourceSharingProtectedResourcesSetting;
 
     /**
      * Constructs a new ResourceAccessControlClient.
      *
      */
-    public ResourceAccessControlClient(ResourceAccessHandler resourceAccessHandler, Set<String> resourceIndices) {
+    public ResourceAccessControlClient(
+        ResourceAccessHandler resourceAccessHandler,
+        ResourcePluginInfo resourcePluginInfo,
+        OpensearchDynamicSetting<List<String>> resourceSharingProtectedResourcesSetting
+    ) {
         this.resourceAccessHandler = resourceAccessHandler;
-        this.resourceIndices = resourceIndices;
+        this.resourcePluginInfo = resourcePluginInfo;
+        this.resourceSharingProtectedResourcesSetting = resourceSharingProtectedResourcesSetting;
     }
 
     /**
@@ -50,7 +56,7 @@ public final class ResourceAccessControlClient implements ResourceSharingClient 
     @Override
     public void verifyAccess(String resourceId, String resourceIndex, String action, ActionListener<Boolean> listener) {
         // following situation will arise when resource is onboarded to framework but not marked as protected
-        if (!resourceIndices.contains(resourceIndex)) {
+        if (!resourcePluginInfo.getResourceIndicesForProtectedTypes().contains(resourceIndex)) {
             LOGGER.warn(
                 "Resource '{}' is onboarded to sharing framework but is not marked as protected. Action {} is allowed.",
                 resourceId,
@@ -63,32 +69,6 @@ public final class ResourceAccessControlClient implements ResourceSharingClient 
     }
 
     /**
-     * Shares a resource with specified users, roles, or backend roles.
-     *
-     * @param resourceId    The ID of the resource to share.
-     * @param resourceIndex The index containing the resource.
-     * @param target        The recipients of the resource, including users, roles, and backend roles and respective access levels.
-     * @param listener      Callback receiving the updated {@link ResourceSharing} document.
-     */
-    @Override
-    public void share(String resourceId, String resourceIndex, ShareWith target, ActionListener<ResourceSharing> listener) {
-        resourceAccessHandler.share(resourceId, resourceIndex, target, listener);
-    }
-
-    /**
-     * Revokes previously granted access to a resource for specific users or roles.
-     *
-     * @param resourceId        The ID of the resource.
-     * @param resourceIndex     The index containing the resource.
-     * @param target            A map of entities whose access is to be revoked.
-     * @param listener          Callback receiving the updated {@link ResourceSharing} document.
-     */
-    @Override
-    public void revoke(String resourceId, String resourceIndex, ShareWith target, ActionListener<ResourceSharing> listener) {
-        resourceAccessHandler.revoke(resourceId, resourceIndex, target, listener);
-    }
-
-    /**
      * Lists all resources the current user has access to within the given index.
      *
      * @param resourceIndex The index to search for accessible resources.
@@ -97,5 +77,15 @@ public final class ResourceAccessControlClient implements ResourceSharingClient 
     @Override
     public void getAccessibleResourceIds(String resourceIndex, ActionListener<Set<String>> listener) {
         resourceAccessHandler.getOwnAndSharedResourceIdsForCurrentUser(resourceIndex, listener);
+    }
+
+    /**
+     * Returns a flag to indicate whether resource-sharing is enabled for resource-type
+     * @param resourceType the type for which resource-sharing status is to be checked
+     * @return true if enabled, false otherwise
+     */
+    @Override
+    public boolean isFeatureEnabledForType(String resourceType) {
+        return resourceSharingProtectedResourcesSetting.getDynamicSettingValue().contains(resourceType);
     }
 }